+++ CHANGELOG.md
@@ -5,6 +5,7 @@ v2.7.7 [unreleased]
 
 ### Bugfixes
 - [#662](https://github.com/hydra-billing/homs/pull/662) Change scope parameter to array.
+- [#664](https://github.com/hydra-billing/homs/pull/664) Fix hide start BP button in case of `candidate_starters` enabled.
 
 v2.7.6 [2022-06-28]
 -------------------